Remaining no-CGAL
